pyproject.toml
[tool.poetry]
name = "shoppyshops"
version = "0.1.0"
description = ""
authors = ["Daryl Antony <daryl@commoncode.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "^3.4.3"
aiohttp = "^3.10.10"
django = "^5.1.2"
django-htmx = "^1.20.0"
reloady = "^0.1.7"
python-dotenv = "^1.0.1"
ipdb = "^0.13.13"
django-livereload-server = "^0.5.1"
daphne = "^4.1.2"
ebaysdk = "^2.2.0"
httpx = "^0.27.2"
pyperclip = "^1.9.0"
pytest = "^8.3.4"
pytest-watch = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
